pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "dcv2nav"
version = "1.0.0"
description = "Discord Components v2 navigation views for Red-DiscordBot cogs"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "ltzmax" }]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.7.1",
]
keywords = ["discord", "redbot", "components", "v2", "pagination", "navigation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
]

[project.urls]
Homepage = "https://github.com/ltzmax/maxcogs-utils"
Repository = "https://github.com/ltzmax/maxcogs-utils"
Issues = "https://github.com/ltzmax/maxcogs-utils/issues"

[tool.setuptools.packages.find]
include = ["maxcogs-utils*"]
